refactor: restruct schedule folder
