+++ src/pages/TourBuilder/Schedule/ActivityBuilder.tsx
@@ -1,5 +1,5 @@
 import { getInitialData } from ".";
-import { Board } from "./TwoColumns/shared/board";
+import { Board } from "./shared/Board";
 
 const ActivityBuilder = () => {
   return <Board initial={getInitialData()} />;
