+++ src/pages/TourBuilder/Schedule/shared/Board.tsx
@@ -11,8 +11,8 @@ import { blockBoardPanningAttr } from "./data-attributes";
 import { CleanupFn } from "@atlaskit/pragmatic-drag-and-drop/dist/types/internal-types";
 import TourBuilderLayout from "@/layouts/TourBuilderLayout";
 import { Card } from "@/components/ui/card";
-import { DestinationCardList } from "./destination-card";
-import { SourceCardList } from "./source-card";
+import { DestinationCardList } from "./DestinationCard";
+import { SourceCardList } from "./SourceCard";
 import { Tour } from "@/types";
 import { Button } from "@/components/ui/button";
 import { Plus } from "lucide-react";
