+++ src/pages/TourBuilder/Schedule/shared/is-safari.ts
